refactor: clean up file and directory structure
