+++ scripts/00_compile_all.sh
