+++ src/libraw_wrapper.cpp
